requirements.txt
certifi==2021.10.8
EasyProcess==0.3
entrypoint2==0.2.4
MouseInfo==0.1.3
mss==6.1.0
Pillow==8.4.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.5
pyperclip==1.8.2
PyRect==0.1.4
pyscreenshot==3.0
PyScreeze==0.1.28
pytweening==1.0.4
six==1.16.0
wincertstore==0.2
